JP-4438 Added attributes, needs further testing

+++ comu/com.energyict.mdc.protocol.api/src/main/java/com/energyict/mdc/protocol/api/device/messages/BaseDeviceMessageFactory.java
@@ -11,6 +11,6 @@ import com.energyict.mdc.common.ApplicationComponent;
  */
 public interface BaseDeviceMessageFactory {
 
-    public DeviceMessage findDeviceMessage(int messageId);
+    public DeviceMessage findDeviceMessage(long messageId);
 
 }
