+++ comu/com.energyict.mdc.protocol.api/src/main/java/com/energyict/mdc/protocol/api/device/offline/OfflineDeviceMessage.java
@@ -28,7 +28,7 @@ public interface OfflineDeviceMessage extends Offline {
     /**
      * @return the ID of the DeviceMessage
      */
-    public int getDeviceMessageId();
+    public long getDeviceMessageId();
 
     /**
      * The ID of the Device owning this DeviceMessage
